Add Android Support
- Update README.md
- Remove window platform
