+++ android/src/main/java/com/shukerullah/lotame/RNLotamePackage.java
@@ -1,5 +1,5 @@
 
-package com.reactlibrary;
+package com.shukerullah.lotame;
 
 import java.util.Arrays;
 import java.util.Collections;
