+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-native-lotame",
-  "version": "1.0.2",
+  "version": "2.0.0",
   "description": "",
   "main": "index.js",
   "scripts": {
@@ -15,8 +15,5 @@
     "lotame"
   ],
   "author": "",
-  "license": "",
-  "peerDependencies": {
-    "react-native": "^0.41.2"
-  }
+  "license": ""
 }
