--- windows/.npmignore
@@ -1,9 +1,0 @@
-
-# Make sure we don't publish build artifacts to NPM
-ARM/
-Debug/
-x64/
-x86/
-bin/
-obj/
-.vs/
