--- windows/RNLotame/RNLotameModule.cs
@@ -1,33 +1,0 @@
-using ReactNative.Bridge;
-using System;
-using System.Collections.Generic;
-using Windows.ApplicationModel.Core;
-using Windows.UI.Core;
-
-namespace Lotame.RNLotame
-{
-    /// <summary>
-    /// A module that allows JS to share data.
-    /// </summary>
-    class RNLotameModule : NativeModuleBase
-    {
-        /// <summary>
-        /// Instantiates the <see cref="RNLotameModule"/>.
-        /// </summary>
-        internal RNLotameModule()
-        {
-
-        }
-
-        /// <summary>
-        /// The name of the native module.
-        /// </summary>
-        public override string Name
-        {
-            get
-            {
-                return "RNLotame";
-            }
-        }
-    }
-}
